requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.35
plotly==5.18.0
matplotlib==3.8.2
requests==2.31.0
scikit-learn==1.4.0
pytest==7.4.4
pytest-cov==4.1.0
flake8==7.0.0
black==24.1.1
